requirements.txt
sqlalchemy
python-dotenv
pandas
pymysql
cryptography